+++ src/docs/SCR_Menu.svelte
@@ -5,58 +5,62 @@
 <div class="scr-menu">
   <h4 class="scr-menu-h4">Introduction</h4>
   <SCR_ROUTER_LINK to={{ name: "rootRoute" }}>
-    <div class="scr-menu-item">Presentation</div>
+    <div class="scr-btn">Presentation</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "installationRoute" }}>
-    <div class="scr-menu-item">Installation</div>
+    <div class="scr-btn">Installation</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "gettingStartedRoute" }}>
-    <div class="scr-menu-item">Getting Started</div>
+    <div class="scr-btn">Getting Started</div>
   </SCR_ROUTER_LINK>
   <h4 class="scr-menu-h4">Configuration</h4>
   <SCR_ROUTER_LINK to={{ name: "configurationOptionsRoute" }}>
-    <div class="scr-menu-item">Configuration Options</div>
+    <div class="scr-btn">Configuration Options</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "configurationGlobalBeforeEnterOptionRoute" }}>
-    <div class="scr-menu-item">Global Before Enter</div>
+    <div class="scr-btn">Global Before Enter</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "configurationOnErrorOptionRoute" }}>
-    <div class="scr-menu-item">Global On Error</div>
+    <div class="scr-btn">Global On Error</div>
   </SCR_ROUTER_LINK>
   <h4 class="scr-menu-h4">Route Object</h4>
   <SCR_ROUTER_LINK to={{ name: "routeObjectOptionsRoute" }}>
-    <div class="scr-menu-item">Properties</div>
+    <div class="scr-btn">Properties</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "routeObjectBeforeEnterRoute" }}>
-    <div class="scr-menu-item">Before Enter</div>
+    <div class="scr-btn">Before Enter</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "routeObjectAfterBeforeEnterRoute" }}>
-    <div class="scr-menu-item">After Before Enter</div>
+    <div class="scr-btn">After Before Enter</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "routeObjectOnErrorRoute" }}>
-    <div class="scr-menu-item">On Error</div>
+    <div class="scr-btn">On Error</div>
   </SCR_ROUTER_LINK>
   <h4 class="scr-menu-h4">Route Component</h4>
   <SCR_ROUTER_LINK to={{ name: "routeComponentPropertiesRoute" }}>
-    <div class="scr-menu-item">Properties</div>
+    <div class="scr-btn">Properties</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "routeComponentComponentsRoute" }}>
-    <div class="scr-menu-item">Components</div>
+    <div class="scr-btn">Components</div>
   </SCR_ROUTER_LINK>
   <h4 class="scr-menu-h4">Navigation</h4>
   <SCR_ROUTER_LINK to={{ name: "navigationRoutingRoute" }}>
-    <div class="scr-menu-item">Routing</div>
+    <div class="scr-btn">Routing</div>
   </SCR_ROUTER_LINK>
   <SCR_ROUTER_LINK to={{ name: "navigationStoreRoute" }}>
-    <div class="scr-menu-item">Store</div>
+    <div class="scr-btn">Store</div>
   </SCR_ROUTER_LINK>
   <h4 class="scr-menu-h4">Route Link</h4>
   <SCR_ROUTER_LINK to={{ name: "routerLinkPropertiesRoute" }}>
-    <div class="scr-menu-item">Properties</div>
+    <div class="scr-btn">Properties</div>
   </SCR_ROUTER_LINK>
   <h4 class="scr-menu-h4">Router Store</h4>
   <SCR_ROUTER_LINK to={{ name: "routerStorePropertiesRoute" }}>
-    <div class="scr-menu-item">Properties</div>
+    <div class="scr-btn">Properties</div>
+  </SCR_ROUTER_LINK>
+  <h4 class="scr-menu-h4">Testing</h4>
+  <SCR_ROUTER_LINK to={{ name: "test1Route" }}>
+    <div class="scr-btn">Test One</div>
   </SCR_ROUTER_LINK>
 </div>
 
@@ -83,25 +87,4 @@
     color: #d43900;
   }
 
-  .scr-menu-item {
-    display: grid;
-    grid-template-columns: 1fr;
-    justify-content: left;
-    align-content: center;
-    width: 100%;
-    padding: 0.5rem;
-    background-color: #ff440009;
-    border-radius: 0.5rem;
-    transition: all linear 0.2s;
-    text-align: center;
-    color: #a82d00;
-    margin-bottom: 0.5rem;
-  }
-
-  .scr-menu-item:hover {
-    background-color: #ff440018;
-    cursor: pointer;
-    font-weight: bolder;
-    color: #d43900;
-  }
 </style>