+++ src/js/store/router.js
@@ -4,7 +4,7 @@ import {
   getStoreState,
   updateStoreKey,
   getStoreKey,
-} from "../functions";
+} from "../helpers/functions";
 import * as lsPlugin from "../../plugins/lsplugin.js";
 import * as lfPlugin from "../../plugins/lfplugin.js";